requirements.txt
# Flask Exam Monitoring System Requirements

# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7

# Computer Vision
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2

# Data Processing
pandas==2.0.3

# Database (optional)
psycopg2-binary==2.9.7

# Utilities
tabulate==0.9.0
python-dateutil==2.8.2

# Development
pytest==7.4.2
pytest-flask==1.2.0